fix: reference schematic graphic assets
