+++ src/schematic/schematic_graphic.ts
@@ -1,5 +1,5 @@
 import { z } from "zod"
-import { getZodPrefixedIdWithDefault } from "src/common"
+import { asset, type Asset, getZodPrefixedIdWithDefault } from "src/common"
 import { expectTypesMatch } from "src/utils/expect-types-match"
 
 export const schematic_graphic = z
@@ -7,21 +7,23 @@ export const schematic_graphic = z
     type: z.literal("schematic_graphic"),
     schematic_graphic_id: getZodPrefixedIdWithDefault("schematic_graphic"),
     schematic_sheet_id: z.string().optional(),
-    svg_content: z.string(),
+    asset,
+    svg_content: z.string().optional(),
   })
-  .describe("Embeds SVG content on a schematic sheet")
+  .describe("References a graphic asset on a schematic sheet")
 
 export type SchematicGraphicInput = z.input<typeof schematic_graphic>
 type InferredSchematicGraphic = z.infer<typeof schematic_graphic>
 
 /**
- * Embeds SVG content on a schematic sheet.
+ * References a graphic asset on a schematic sheet.
  */
 export interface SchematicGraphic {
   type: "schematic_graphic"
   schematic_graphic_id: string
   schematic_sheet_id?: string
-  svg_content: string
+  asset: Asset
+  svg_content?: string
 }
 
 expectTypesMatch<SchematicGraphic, InferredSchematicGraphic>(true)
